Working mockup
A python gtk app that shows two labels, with a thick border.

In both label a base color, the same for both, is used for the border,
while a modification of the same base color is used for the background.

The first label is the target, with the base color defined as sass
variable, and the modified background color obtained with sass
functions.

The second label has the base color defined as CSS (gtk) variable and the
modified background one is obtained with a CSS function.

The two labels are expected to have the very same colors.